moved manager implementation to store_manager
